Initial commit of the Catalyst port
This is very incomplete. It includes lots of schema fixes for compilation issues
among other problems. I do not have a single working action yet, but I'm close
on the journal action. The journal_items methods need to be added to make it go
and I'm trying to decide what that will look like..