Fixing up time zone support in the Catalyst version
Up till now, all dates were left in UTC. With this patch, the time zone set in
the user's preferences is used as the time zone for displaying all dates and
times. Only the server code actually knows about and understands this, the
schema code handles time zones, but requires the caller to explain what the time
zone is.

There is one important problem right now, which is that the JournalDay objects
have a datestamp that is stored in UTC. However, this does not match reality.
The reality is that the stored date is actually a floating time zone date
corresponding to whichever time zone the user prefers. I'm not sure how best to
correct this yet. The notion of a journal day might not be necessary anymore anyway
(it's original purpose being to segment out days in a way that was conveniently
accessed from Jifty::DBI, which is no longer in use).