Starting to fix broken SOC regarding journal item generation
Currently, the list in the journal or when viewing tasks and tags is generated
by calling a method on the model result classes. While convenient, this means
that the result classes are now entangled with the view classes for the server,
which is very naughty. I'm moving the implementation into a special view. The
model gives us a blueprint that tells us the general ordering and such, but does
not dictate the content of the view.